# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AstraWatch Unified API",
    "description": "Official OpenAPI 3.0 specification for AstraWatch autonomous telemetry, anomaly detection, incident orchestration, and K8s self-healing microservices.",
    "version": "1.0.0",
    "contact": {
      "name": "AstraWatch Engineering",
      "url": "https://astrawatch.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Collector Service (:8080)"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Analyzer Service (:8000)"
    },
    {
      "url": "http://localhost:8082/api/v1",
      "description": "Orchestrator Service (:8082)"
    },
    {
      "url": "http://localhost:8081/healthz",
      "description": "K8s Operator Service (:8081 healthz)"
    },
    {
      "url": "http://localhost:8084",
      "description": "Realtime Gateway (:8084)"
    },
    {
      "url": "http://localhost:8085/api/v1",
      "description": "Payment Service (:8085)"
    }
  ],
...
  "paths": {
...
    "/api/v1/slo/:serviceId": {
      "get": {
        "tags": [
          "Orchestrator: SLO"
        ],
        "summary": "Get SLO Metrics",
        "description": "Returns SLO metrics: current attainment percentage, error budget remaining, and burn rate.",
        "operationId": "os1",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "description": "Service ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "serviceId": "svc-001",
                  "sloTarget": 99.9,
                  "currentAttainment": 99.7,
                  "errorBudgetRemaining": 0.12,
                  "burnRate": 2.4
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/slo": {
      "post": {
        "tags": [
          "Orchestrator: SLO"
        ],
        "summary": "Define SLO",
        "description": "Defines a new Service Level Objective for a service.",
        "operationId": "os2",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "slo-001",
                  "serviceId": "svc-001",
                  "targetPercentage": 99.9,
                  "windowDays": 30
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "SLO definition",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "serviceId": "svc-001",
                "metric": "availability",
                "targetPercentage": 99.9,
                "windowDays": 30
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer JWT token issued by POST /api/v1/auth/login"
      }
    },
...
  }
}